moving Forms to Atoms
