+++ arlo-client/src/components/Audit/SelectBallotsToAudit.tsx
@@ -22,10 +22,10 @@ import { Link } from 'react-router-dom'
 import FormSection, {
   FormSectionDescription,
   FormSectionLabel,
-} from '../Form/FormSection'
-import FormWrapper from '../Form/FormWrapper'
-import FormButton from '../Form/FormButton'
-import FormButtonBar from '../Form/FormButtonBar'
+} from '../Atoms/Form/FormSection'
+import FormWrapper from '../Atoms/Form/FormWrapper'
+import FormButton from '../Atoms/Form/FormButton'
+import FormButtonBar from '../Atoms/Form/FormButtonBar'
 import {
   IJurisdiction,
   IAudit,
@@ -34,9 +34,9 @@ import {
   IErrorResponse,
 } from '../../types'
 import { api, testNumber, checkAndToast, poll } from '../utilities'
-import { generateOptions, ErrorLabel } from '../Form/_helpers'
+import { generateOptions, ErrorLabel } from '../Atoms/Form/_helpers'
 import H2Title from '../Atoms/H2Title'
-import FormField from '../Form/FormField'
+import FormField from '../Atoms/Form/FormField'
 import number, { parse as parseNumber } from '../../utils/number-schema'
 import { formattedUpTo } from '../../utils/indexes'
