+++ arlo-client/src/components/Audit/Setup/Participants/index.tsx
@@ -6,14 +6,16 @@ import { toast } from 'react-toastify'
 import styled from 'styled-components'
 import { HTMLSelect, Spinner, FileInput } from '@blueprintjs/core'
 import { IErrorResponse } from '../../../../types'
-import FormWrapper from '../../../Form/FormWrapper'
-import FormButtonBar from '../../../Form/FormButtonBar'
-import FormButton from '../../../Form/FormButton'
+import FormWrapper from '../../../Atoms/Form/FormWrapper'
+import FormButtonBar from '../../../Atoms/Form/FormButtonBar'
+import FormButton from '../../../Atoms/Form/FormButton'
 import { IValues } from './types'
 import labelValueStates from './states'
 import schema from './schema'
-import { ErrorLabel } from '../../../Form/_helpers'
-import FormSection, { FormSectionDescription } from '../../../Form/FormSection'
+import { ErrorLabel } from '../../../Atoms/Form/_helpers'
+import FormSection, {
+  FormSectionDescription,
+} from '../../../Atoms/Form/FormSection'
 import { api, checkAndToast } from '../../../utilities'
 import { ISidebarMenuItem } from '../../../Atoms/Sidebar'
 import useAuditSettings from '../useAuditSettings'
