+++ arlo-client/src/components/Audit/Setup/Settings/index.tsx
@@ -2,16 +2,16 @@ import React from 'react'
 import { useParams } from 'react-router-dom'
 import { Formik, FormikProps, Form, Field, ErrorMessage } from 'formik'
 import { RadioGroup, Radio, Spinner } from '@blueprintjs/core'
-import FormButtonBar from '../../../Form/FormButtonBar'
-import FormButton from '../../../Form/FormButton'
+import FormButtonBar from '../../../Atoms/Form/FormButtonBar'
+import FormButton from '../../../Atoms/Form/FormButton'
 import { ISidebarMenuItem } from '../../../Atoms/Sidebar'
 import { IValues } from './types'
-import FormWrapper from '../../../Form/FormWrapper'
-import FormSection from '../../../Form/FormSection'
+import FormWrapper from '../../../Atoms/Form/FormWrapper'
+import FormSection from '../../../Atoms/Form/FormSection'
 import { Select } from '../../EstimateSampleSize'
-import { generateOptions, ErrorLabel } from '../../../Form/_helpers'
+import { generateOptions, ErrorLabel } from '../../../Atoms/Form/_helpers'
 import { parse as parseNumber } from '../../../../utils/number-schema'
-import FormField from '../../../Form/FormField'
+import FormField from '../../../Atoms/Form/FormField'
 import schema from './schema'
 import useAuditSettings from '../useAuditSettings'
